MdnsLite recursive DNS no longer needed
Now that we're using Erlang/OTP 24.1, it's possible to fall back to the
regular DNS resolver for regular DNS lookups. This should work better
and it also means that regular DNS lookups will still work if MdnsLite
goes down.

The OTP 24.1 fix that makes this possible is the one where you can bind
a UDP socket first to localhost and then rebind it to a network
interface. The fix is a workaround to a Linux issue with doing this.