Save notebooks to /data/livebook
This removes the "s" so that the data directory is the same as the
application name. This also is more consistent with the use of
"Livebook" to refer to the app and "notebook" or "code notebook" to
refer to the files.

This will certainly be an annoying update, but better late than never.